pyproject.toml
[tool.poetry]
name = "drippy"
version = "0.1.0"
description = "Silly scripts to control your digital ocean droplets."
authors = ["Matthew Karas <1146938+lucidguppy@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
click = "^8.1.3"
pexpect = "^4.8.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
